Kenny's Work v1
I added footsteps, forest ambience, and the battle theme. I created an audio mixer and leveled the audio.

Footsteps: I edited the TestMove.cs script adding one if statement; and an audio source for the footstep.

Battle Theme: I Added an audio source for the Battle Theme. This is in the BattleScene to play on wake.