Turn Order WIP3
The stats were already a class. Derp. This changes the stats back to the way it was. Minor changes to Attack.cs, PlayerCharacter.prefab. BattleReady has received an overhaul preparing for the work on turn order (which has already started, but I got stuck and posted to the Unity forums).