Untested: Beet attack + a key fundamental change
This code is currently untested, but I realized that when "Beet attacks!" was a message in the console, it was only because it was attached to the player character. This means it wouldn't be able to access the Beet's stats; a condition was added to Fight() that asks if it's my turn (GO whose indexNo it is == me?) before executing any attack function. I may move the battle initialization to the BattleReady script.