Overworld Map is In!
Brought the map into the project, but it overwrote the inventory test scene, and I have no idea why. Tried to reimport the inventory test scene (all the prefabs, etc work, right?) but then it wanted to overwrite the newly written MapOverworld scene (despite not being named such in the original...)  ~0.5 hours.