Attack Canvas and UI WIP3
Clicking on a button for an attack/move selects that attack. Upon ResetAttacks(), the menu and move choice (integer-based) are set inactive and to 0.
If it is the chef's turn, "Attack" comes up, whereas it reads "Move" for the DG's turn. When the turn resets to 0 (the first to move, usually DG), "Attack" is there, but if you click "Attack" it becomes "Move" and then if you click "Move" it expands the body contents of the Move menu as expected.
Targeting is still performed using the 5/6/8 keys, but I intend to change this to clicking on the target.