Attack UI
Fixed a bug where you selected the same move many times and used up all your magic. No trade-backsies anymore if you want to change your move, either. By changing the Battle Camera to be tagged "MainCamera" like the Overworld Camera, I'm able to use a raycast to determine the GameObject clicked on and make it the target if it is in the list of combatants (according to WhoseTurn's "order" list).